+++ Utils.cpp
@@ -1,0 +1,20 @@
+#include <cstdlib> // For exit() and EXIT_FAILURE
+#include <cstring> // For memset
+#include <iostream>
+#include <netinet/in.h> // For sockaddr_in
+#include <sys/socket.h> // For socket functions
+#include <unistd.h>     // For read, write, and close
+#include <fstream>
+#include <sstream>
+#include "include/webserv.hpp"
+
+std::string readHtml(const std::string& filePath) {
+    std::ifstream file(filePath);
+    if (!file.is_open()) {
+        std::cerr << "Could not open file: " << filePath << std::endl;
+        return "";
+    }
+    std::stringstream buffer;
+    buffer << file.rdbuf();
+    return buffer.str();
+}
