+++ main.cpp
@@ -9,115 +9,6 @@
 #include "include/webserv.hpp"
 
 const int PORT = 8080;
-const int BUFFER_SIZE = 1024;
-
-std::string readHtml(const std::string& filePath) {
-    std::ifstream file(filePath);
-    if (!file.is_open()) {
-        std::cerr << "Could not open file: " << filePath << std::endl;
-        return "";
-    }
-    std::stringstream buffer;
-    buffer << file.rdbuf();
-    return buffer.str();
-}
-
-std::string handleHomePage() {
-    std::string htmlContent = readHtml("./html/home.html");
-    std::string httpResponse = "HTTP/1.1 200 OK\nContent-Type: text/html\n" +
-                               std::string("Content-Length: ") + std::to_string(htmlContent.length()) + "\n\n" +
-                               htmlContent;
-    std::cout << "------------------Home page returned from handleHomePage()-------------------" << std::endl;
-    return httpResponse;
-}
-
-std::string handleCGIRequest(const char* argv[], const char* envp[]) {
-    int pipefd[2];
-    if (pipe(pipefd) == -1) {
-        perror("pipe failed");
-        exit(EXIT_FAILURE);
-    }
-
-    pid_t pid = fork();
-    if (pid == -1) {
-        perror("fork failed");
-        exit(EXIT_FAILURE);
-    } else if (pid == 0) {
-        close(pipefd[0]);
-        dup2(pipefd[1], STDOUT_FILENO);
-        close(pipefd[1]);
-
-    //annoying casts to make execve happy
-    if (execve(argv[0], const_cast<char* const*>(argv), const_cast<char* const*>(envp)) == -1) {
-        perror("execve failed");
-        exit(EXIT_FAILURE);
-    }
-    } else {
-        close(pipefd[1]);
-
-        std::string cgiOutput;
-        char readBuffer[256];
-        ssize_t bytesRead;
-        while ((bytesRead = read(pipefd[0], readBuffer, sizeof(readBuffer) - 1)) > 0) {
-            readBuffer[bytesRead] = '\0';
-            cgiOutput += readBuffer;
-        }
-        close(pipefd[0]);
-
-        int status;
-        waitpid(pid, &status, 0);
-
-        // Construct the HTTP response
-        std::string httpResponse = "HTTP/1.1 200 OK\nContent-Type: text/html\n" +
-                                   std::string("Content-Length: ") + std::to_string(cgiOutput.length()) + "\n\n" +
-                                   cgiOutput;
-        
-        std::cout << "------------------CGI output prepared-------------------" << std::endl;
-        return httpResponse;
-    }
-
-    //line necessary to compile
-    return "HTTP/1.1 500 Internal Server Error\nContent-Length: 0\n\n";
-}
-
-
-
-std::string handleNotFound(void) {
-    std::string response = "HTTP/1.1 404 Not Found\r\nContent-Length: 0\r\n\r\n";
-    // write(socket, response.c_str(), response.size());
-    std::cout << "------------------404 Not Found sent-------------------" << std::endl;
-
-    return response;
-}
-
-
-    // Determine the type of request and call the appropriate handler
-void handleConnection(int socket) {
-    char buffer[BUFFER_SIZE] = {0};
-    long valread = read(socket, buffer, BUFFER_SIZE);
-    if (valread < 0) {
-        perror("In read");
-        exit(EXIT_FAILURE);
-    }
-    std::cout << "Received HTTP request: " << std::endl << buffer << std::endl;
-
-    //test to execute the python script (see: https://www.tutorialspoint.com/python/python_cgi_programming.htm)
-    // const char* argv[] = { "./cgi-bin/hello_py.cgi", NULL };
-    const char* argv[] = { "./cgi-bin/hello.cgi", NULL };
-    const char* envp[] = { "QUERY_STRING=Hello from C++ CGI!", NULL };
-
-    std::string response;
-    if (strstr(buffer, "GET / HTTP/1.1") || strstr(buffer, "GET /home HTTP/1.1")) {
-        response = handleHomePage();
-    } else if (strstr(buffer, "GET /hello HTTP/1.1")) {
-        response = handleCGIRequest(argv, envp);
-    } else {
-        response = handleNotFound();
-    }
-
-    write(socket, response.c_str(), response.size());
-    close(socket);
-    }
 
 int main()
 {
